Update run.py
I wasnt able to test it with recent set of changes checked in, but I have tried with previous set of commands and was able to generate .txt file which contains extracted IPC and a PNG with graphical analysis of the commands ran. To test it out you'd just need to replace these cmds with newly added cmds and that should do! Lmk  if there are any questions or concerns!